// tmp_app/Api.ts
export interface input$api$auth$update_password {
    password: string, old_password: string
}
// ...
export interface output$api$auth$update_password {
    f_name: string, l_name: string, username: string, email: string, channel: string
}